land: Add shim to report old API level
In Pie, Google added checks to prevent destroying of the same mutexes
consecutively, and aborts the program on fatal call.

The check was triggered by some camera component, which destroyed a
mutex consecutively by adding the destroy instruction in a loop.

However, they have also allowed apps with API level below
__ANDROID_API_P__ to skip abortion and continue normally. Make use of
this fact and report api level < 28 for the whole camera stack.

Change-Id: I4f8542a496dc25c53f8208a647cc0d6acc8662eb