land: Add pthread_mutex_destroy hack for camera
In Pie, Google added checks to prevent destroying of the same mutexes
consecutively, and aborts the program on fortify fatal call.

The check is triggered when tearing down the ISP session, which destroys
the same mutex consecutively by adding the destroy instruction in a loop.
This causes the whole daemon to crash and fail.

(Specifically, the violating instructions were from isp_module_stop_session
and when isp_module_start_session fails to create a thread.)

Change-Id: I4f8542a496dc25c53f8208a647cc0d6acc8662eb